move updatePoliciesStatus() to leader.go where it is used
