+++ internal/k8s/leader.go
@@ -2,10 +2,13 @@ package k8s
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"time"
 
 	"github.com/golang/glog"
+	conf_v1 "github.com/nginxinc/kubernetes-ingress/pkg/apis/configuration/v1"
+	"github.com/nginxinc/kubernetes-ingress/pkg/apis/configuration/validation"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -110,4 +113,34 @@ func (lbc *LoadBalancerController) addLeaderHandler(leaderHandler leaderelection
 	if err != nil {
 		glog.V(3).Infof("Error starting LeaderElection: %v", err)
 	}
+}
+
+func (lbc *LoadBalancerController) updatePoliciesStatus() error {
+	var allErrs []error
+	for _, nsi := range lbc.namespacedInformers {
+		for _, obj := range nsi.policyLister.List() {
+			pol := obj.(*conf_v1.Policy)
+
+			err := validation.ValidatePolicy(pol, lbc.isNginxPlus, lbc.enableOIDC, lbc.appProtectEnabled)
+			if err != nil {
+				msg := fmt.Sprintf("Policy %v/%v is invalid and was rejected: %v", pol.Namespace, pol.Name, err)
+				err = lbc.statusUpdater.UpdatePolicyStatus(pol, conf_v1.StateInvalid, "Rejected", msg)
+				if err != nil {
+					allErrs = append(allErrs, err)
+				}
+			} else {
+				msg := fmt.Sprintf("Policy %v/%v was added or updated", pol.Namespace, pol.Name)
+				err = lbc.statusUpdater.UpdatePolicyStatus(pol, conf_v1.StateValid, "AddedOrUpdated", msg)
+				if err != nil {
+					allErrs = append(allErrs, err)
+				}
+			}
+		}
+	}
+
+	if len(allErrs) != 0 {
+		return fmt.Errorf("not all Policies statuses were updated: %v", allErrs)
+	}
+
+	return nil
 }
